made a CountsContext to handle quiz counts changes
